requirements.txt
fastapi>=0.111.0
pydantic-settings>=2.0.0
uvicorn[standard]>=0.29.0
